RESTRUCTURED AUTHENTICATION WINDOW LOADING - Moved authenticate() under home_view.dart
