+++ lib/main.dart
@@ -17,24 +17,9 @@ void main() async {
   await Hive.initFlutter();
   await Hive.openBox('flutter_notes_app');
 
-  var localAuth = LocalAuthentication();
-
-  SharedPreferences prefs = await SharedPreferences.getInstance();
-
-  if (prefs.getBool('auth') == true) {
-    var authenticate = await localAuth.authenticate(
-      localizedReason: 'Authenticate to access Notes',
-      useErrorDialogs: true,
-      stickyAuth: true,
-    );
-    if (authenticate == true) {
-      await SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp, DeviceOrientation.portraitDown])
-          .then((_) {
-        runApp(ProviderScope(child: MyApp()));
-      });
-    } else
-      exit(0);
-  }
+  await SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp, DeviceOrientation.portraitDown]).then((_) {
+    runApp(ProviderScope(child: MyApp()));
+  });
 }
 
 class MyApp extends StatelessWidget {
